requirements.txt
matplotlib==3.7.1
numpy==1.24.3
pandas==2.0.2
python-binance==1.0.17
python-dotenv==1.0.0
statsmodels==0.14.0
scipy~=1.10.1